pyproject.toml
[project]
name = "tg-summarizer"
version = "0.1.0"
description = "Telegram bot that collects messages from public channels, filters them based on topics, summarizes the relevant content, and delivers reports to the user."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.1.0",
    "python-telegram-bot[job-queue]>=22.1",
    "sqlalchemy>=2.0.41",
    "telethon>=1.40.0",
    "sentence-transformers>=4.1.0",
    "torch>=2.7.1",
    "hf-xet>=1.1.5",
    "alembic>=1.16.2",
    "qdrant-client>=1.7.0",
    "google-genai>=1.25.0",
    "openai>=1.97.0",
    "pytz>=2025.2",
    "python-json-logger>=3.3.0",
]

[project.scripts]
tg-bot = "bot.main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
